api/requirements.txt
torch==2.7.0
torchvision==0.22.0
pandas==2.2.3
opencv-python==4.8.0.76
tqdm==4.65.0
numpy==1.24.4
pillow==9.5.0
python-dotenv==0.21.0
fastapi==0.95.0
uvicorn==0.22.0